+++ src/main/java/io/github/codesmashing/programming_project_aquafin_case2/programming_project_aquafin_case2/controllers/APIController.java
@@ -1,100 +1,170 @@
 package io.github.codesmashing.programming_project_aquafin_case2.programming_project_aquafin_case2.controllers;
 
 import java.time.LocalDate;
-import java.time.format.TextStyle;
-import java.util.Locale;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import io.github.codesmashing.programming_project_aquafin_case2.programming_project_aquafin_case2.models.Flood;
-import io.github.codesmashing.programming_project_aquafin_case2.programming_project_aquafin_case2.models.Month;
-import io.github.codesmashing.programming_project_aquafin_case2.programming_project_aquafin_case2.models.DAO.FloodDAO;
-import io.github.codesmashing.programming_project_aquafin_case2.programming_project_aquafin_case2.models.DAO.MonthDAO;
-import io.github.codesmashing.programming_project_aquafin_case2.programming_project_aquafin_case2.models.DAO.SeasonDAO;
+import io.github.codesmashing.programming_project_aquafin_case2.programming_project_aquafin_case2.models.FloodRisk;
+import io.github.codesmashing.programming_project_aquafin_case2.programming_project_aquafin_case2.models.PrecipitationData;
+import io.github.codesmashing.programming_project_aquafin_case2.programming_project_aquafin_case2.models.Region;
+import io.github.codesmashing.programming_project_aquafin_case2.programming_project_aquafin_case2.models.DAO.FloodRiskDAO;
+import io.github.codesmashing.programming_project_aquafin_case2.programming_project_aquafin_case2.models.DAO.PrecipitationDataDAO;
+import io.github.codesmashing.programming_project_aquafin_case2.programming_project_aquafin_case2.models.DAO.RegionDAO;
 
 @RestController
 @RequestMapping("/api")
 public class APIController {
-	private final FloodDAO floodDAO;
-	private final SeasonDAO seasonDAO;
-	private final MonthDAO monthDAO;
+	private final FloodRiskDAO floodRiskDAO;
+	private final RegionDAO regionDAO;
+	private final PrecipitationDataDAO precipitationDataDAO;
 
 	@Autowired
-	public APIController(SeasonDAO seasonDAO, FloodDAO floodDAO, MonthDAO monthDAO) {
-		this.seasonDAO = seasonDAO;
-		this.floodDAO = floodDAO;
-		this.monthDAO = monthDAO;
+	public APIController(RegionDAO regionDAO, FloodRiskDAO floodRiskDAO, PrecipitationDataDAO precipitationDataDAO) {
+		this.regionDAO = regionDAO;
+		this.floodRiskDAO = floodRiskDAO;
+		this.precipitationDataDAO = precipitationDataDAO;
 	}
 
-	@GetMapping({ "", "/" })
-	public ResponseEntity<?> getCurrentMonth() {
-		LocalDate currentDate = LocalDate.now();
-		String currentMonth = currentDate.getMonth().getDisplayName(TextStyle.FULL, Locale.forLanguageTag("nl"));
-		Optional<Flood> floodOptional = floodDAO.findById(currentDate.getYear());
-
-		if (floodOptional.isEmpty()) {
-			return ResponseEntity.status(HttpStatus.NOT_FOUND)
-					.body("Flood not found for current year: " + currentDate.getYear());
-		}
-
-		Optional<Month> monthOptional = monthDAO.findByFloodIdAndName(floodOptional.get(), currentMonth);
-		if (monthOptional.isEmpty()) {
-			return ResponseEntity.status(HttpStatus.NOT_FOUND)
-					.body("Month not found for year " + floodOptional.get().getYear() + ": " + currentMonth);
-		}
-
-		return ResponseEntity.ok(monthOptional.get());
+	@GetMapping({ "/regions" })
+	public ResponseEntity<Iterable<Region>> getRegions() {
+		return ResponseEntity.ok(regionDAO.findAll());
 	}
 
-	@GetMapping({ "{year}", "/{year}" })
-	public ResponseEntity<?> getYear(@PathVariable("year") Integer year) {
-		Optional<Flood> floodOptional = floodDAO.findById(year);
-		if (floodOptional.isEmpty()) {
-			return ResponseEntity.status(HttpStatus.NOT_FOUND)
-					.body("Flood not found: " + year);
-		}
-
-		return ResponseEntity.ok(floodOptional.get());
+	@GetMapping({ "/precipitation/all" })
+	public ResponseEntity<Iterable<PrecipitationData>> getPrecipitations() {
+		return ResponseEntity.ok(precipitationDataDAO.findAll());
 	}
 
-	@GetMapping({ "{year}/{month}", "/{year}/{month}/" })
-	public ResponseEntity<?> getSpecificMonth(
-			@PathVariable("year") Integer year,
-			@PathVariable("month") String month) {
-		Optional<Flood> floodOptional = floodDAO.findById(year);
-		if (floodOptional.isEmpty()) {
-			return ResponseEntity.status(HttpStatus.NOT_FOUND)
-					.body("Flood not found: " + year);
+	@GetMapping({ "/precipitation" })
+	public ResponseEntity<?> getPrecipitationPerRegionPerMonth(
+			@RequestParam(value = "region_id", required = false) Integer id,
+			@RequestParam(value = "region_name", required = false) String name,
+			@RequestParam("date") String dateString) {
+		LocalDate date;
+
+		try {
+			date = LocalDate.parse(dateString);
+		} catch (Exception e) {
+			return ResponseEntity.status(HttpStatus.NOT_ACCEPTABLE)
+					.body("The date was not in a correct format: " + dateString);
 		}
 
-		Boolean isValidMonth = isValidMonth(month);
-		if (!isValidMonth) {
-			return ResponseEntity.status(HttpStatus.BAD_REQUEST)
-					.body("Invalid month format. Please use full Dutch month names.");
+		Optional<Region> regionOptional = Optional.empty();
+
+		if (id != null) {
+			regionOptional = regionDAO.findById(id);
 		}
 
-		Optional<Month> monthOptional = monthDAO.findByFloodIdAndName(floodOptional.get(), month);
-		if (monthOptional.isEmpty()) {
-			return ResponseEntity.status(HttpStatus.NOT_FOUND)
-					.body("Month not found for year " + year + ": " + month);
+		if (regionOptional.isEmpty() && name != null) {
+			regionOptional = regionDAO.findByRegionName(name);
 		}
-		return ResponseEntity.ok(monthOptional.get());
+
+		if (regionOptional.isEmpty()) {
+			return ResponseEntity.status(HttpStatus.NOT_FOUND)
+					.body("Region not found by name: " + name + ". Or by ID: " + id);
+		}
+
+		Optional<PrecipitationData> precipitationOptional = precipitationDataDAO
+				.findByRegionAndDateOccurence(regionOptional.get(), date);
+		if (precipitationOptional.isEmpty()) {
+			return ResponseEntity.status(HttpStatus.NOT_FOUND)
+					.body("Precipitation not found for region " + regionOptional.get().getRegionName() + ": " + date);
+		}
+
+		return ResponseEntity.ok(precipitationOptional.get());
 	}
 
-	private Boolean isValidMonth(String inputMonth) {
-		for (java.time.Month month : java.time.Month.values()) {
-			// Check if the input matches the full Dutch name of the month
-			if (month.getDisplayName(TextStyle.FULL, Locale.forLanguageTag("nl")).equalsIgnoreCase(inputMonth)) {
-				return true;
-			}
+	@GetMapping({ "/flood-risks" })
+	public ResponseEntity<?> getFloodRiskPerRegionPerMonth(
+			@RequestParam(value = "region_id", required = false) Integer id,
+			@RequestParam(value = "region_name", required = false) String name,
+			@RequestParam("date") String dateString) {
+		LocalDate date;
+
+		try {
+			date = LocalDate.parse(dateString);
+		} catch (Exception e) {
+			return ResponseEntity.status(HttpStatus.NOT_ACCEPTABLE)
+					.body("The date was not in a correct format: " + dateString);
 		}
-		return false;
+
+		Optional<Region> regionOptional = Optional.empty();
+
+		if (id != null) {
+			regionOptional = regionDAO.findById(id);
+		}
+
+		if (regionOptional.isEmpty() && name != null) {
+			regionOptional = regionDAO.findByRegionName(name);
+		}
+
+		if (regionOptional.isEmpty()) {
+			return ResponseEntity.status(HttpStatus.NOT_FOUND)
+					.body("Region not found by name: " + name + ". Or by ID: " + id);
+		}
+
+		Optional<FloodRisk> floodRiskOptional = floodRiskDAO
+				.findByRegionAndDateOccurence(regionOptional.get(), date);
+		if (floodRiskOptional.isEmpty()) {
+			return ResponseEntity.status(HttpStatus.NOT_FOUND)
+					.body("Flood risk not found for region " + regionOptional.get().getRegionName() + ": " + date);
+		}
+
+		return ResponseEntity.ok(floodRiskOptional.get());
 	}
+
+	@PostMapping({ "/precipitation" })
+	public ResponseEntity<?> addPrecipitation(
+			@RequestParam(value = "region_id", required = false) Integer id,
+			@RequestParam(name = "region_name", required = false) String name,
+			@RequestParam("date") @DateTimeFormat(pattern = "dd/MM/yyyy") String dateString,
+			@RequestParam("rainfall") Double rainfall,
+			@RequestParam(value = "type") Integer dataType,
+			@RequestParam(value = "source", required = false) String source) {
+		LocalDate date;
+
+		try {
+			date = LocalDate.parse(dateString);
+		} catch (Exception e) {
+			return ResponseEntity.status(HttpStatus.NOT_ACCEPTABLE)
+					.body("The date was not in a correct format: " + dateString);
+		}
+
+		Optional<Region> regionOptional = Optional.empty();
+
+		if (id != null) {
+			regionOptional = regionDAO.findById(id);
+		}
+
+		if (regionOptional.isEmpty() && name != null) {
+			regionOptional = regionDAO.findByRegionName(name);
+		}
+
+		if (regionOptional.isEmpty()) {
+			return ResponseEntity.status(HttpStatus.NOT_FOUND)
+					.body("Region not found by name: " + name + ". Or by ID: " + id);
+		}
+
+		PrecipitationData newPrecipitationData = new PrecipitationData();
+		newPrecipitationData.setDataType(dataType);
+		newPrecipitationData.setDateOccurence(date);
+		newPrecipitationData.setQuantityInMm(rainfall);
+		newPrecipitationData.setRegion(regionOptional.get());
+		newPrecipitationData.setSource(source);
+
+		precipitationDataDAO.save(newPrecipitationData);
+
+		return ResponseEntity.ok("Event created successfully");
+	}
+
+	// (optioneel) POST /flood-risks – handmatig risico toevoegen
 }